Create a "genome file" that bedtools jaccard needs.
Otherwise it crashes if a sequence from B is not in A.

There is obvious repetition of the same computation here.  It is not
satisfying but I need a short term fix.  Hopefully it will put
the genome sequence in the node's memory cache, which will be useful
in the final commands.